+++ elu/src/elu.rs
@@ -1,0 +1,7 @@
+// elu: a clone of muri, assembler for rung/the Nga VM
+
+pub mod elu;
+    pub fn stub() {
+        println!("hello");
+    }
+}
