+++ rung_vm/Cargo.toml
@@ -1,8 +1,9 @@
 [package]
-name = "rung"
+name = "rung_vm"
 version = "0.1.0"
 authors = ["Aidan <calopter@gmail.com>"]
 
+[lib]
+
 [dependencies]
-rung_vm = {path = "rung_vm"}
-elu = {path = "elu"}
+byteorder = "1"
