+++ src/main.rs
@@ -1,0 +1,5 @@
+extern crate rung_vm;
+
+fn main() {
+    let vm = rung_vm::make_vm();
+}
